finally fixed house template generation (a lot less random!)
removed a lot of the random chance involved with the room splitting and generation because it meant that often times a room split would fail. now, the stuff like variation is still random, but it goes through all possibilities (for splits) before deciding whether or not its impossible. the only random chance here is when the room amount ends up less than the template because the house physically cannot be split further, so it re-rolls for a new house and tries again. makes house generation way faster. working on adding furniture in a scriptable objects kinda way (that way its super easy to make new house templates and such, so that minimal coding is involved for a new house template)